requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asgiref==3.10.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.3
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cron_descriptor==2.0.6
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.8
django-celery-beat==2.8.1
django-timezone-field==7.1
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
-e git+https://github.com/salute-developers/GigaAM.git@6a8b511f753670ed38af6529bb89bbdc2191ba6a#egg=gigaam
google-auth==2.43.0
google-genai==1.51.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.11
imap-tools==1.11.0
importlib_metadata==8.7.0
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
kombu==5.5.4
lightning==2.5.6
lightning-utilities==0.15.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannoteai-sdk==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytorch-lightning==2.5.6
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.12.0
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
speechbrain==1.0.3
SQLAlchemy==2.0.44
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.5.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchcodec==0.8.1
torchmetrics==1.8.2
tqdm==4.67.1
triton==3.1.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
zipp==3.23.0
